% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewiring_create_module_summary.R
\name{createModuleSummary}
\alias{createModuleSummary}
\title{Create module summary}
\usage{
createModuleSummary(
  ObjectList,
  modmeth = "VBSR",
  numclus = 1,
  supertype = "refined",
  numdataset = 1
)
}
\arguments{
\item{ObjectList}{Output from \code{preparerewiring()}containing some required parameters.}

\item{modmeth}{Method that uses to generate the supermodule (default:"VBSR").}

\item{numclus}{Integer specifying the cluster detected from 'runrewiring()' (default:1).}

\item{supertype}{Secification of the type of supermodule, "refined" or "raw", (default:"refined")}

\item{numdataset}{Integer specifying the dataset used, 1 when there is only one dataset (default:1).}
}
\value{
htmlinfo the index page to the module summary.
}
\description{
Create module summary for supermodule after the rewiring step.
}
\examples{

## Lets assume that we have already generated the ObjectList, we will load it from
## the folder containing the examples files. After running the 'runrewiring()', running this
## function will generate a module summary based on the ouput from 'runrewiring()'.

## To create html summary on previously run supermodel in new data, we will have
## generated a new ObjectList using the new data by running ‘preparerewiring()’.
##  Running the #' ## ‘createModuleSummary()’ with the new ObjectList will create
## a module summary based on the new dataset but with the gene module from the
## 'runrewiring()' results.

objectlist <- readRDS(file=paste0(system.file('extdata',package='TraRe'),
                      '/prepared_rewiring_example.rds'))

## We are going to create the folder containing
## the graphs, reports, etc, and then we are deleting it.
## If you want to keep it, do not run the last line.

## We are modifying output directory for this example.
objectlist$outdir <- paste(getwd(),'examplefolder',sep='/')

## runrewiring(ObjectList = objectlist)
##createModuleSummary(ObjectList = objectlist, modmeth = "VBSR", numclus = 1,
##                    supertype = "refined", numdataset = 1)
##unlink(objectlist$outdir,recursive = TRUE)


}
