% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewiring_stats_functions.R
\name{dave_test}
\alias{dave_test}
\alias{dave_test_pair_detail}
\title{Generate stats from the rewiring test.}
\usage{
dave_test(x, grp, perm = 500)

dave_test_pair_detail(x, grp, perm = 500)
}
\arguments{
\item{x}{data matrix containing subjects as rows and genes as columns.}

\item{grp}{indicator of group membership}

\item{perm}{number of permutations for the test.}
}
\description{
When performing the rewiring test, some stats must be generated in order to proceed with the
rewiring method. \code{dave_test()} generates the correlations to work with. \code{dave_test_pair_detail()}
can produce the correlation plots.
}
\examples{
\dontrun{
  example here
}

}
