% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_generation.R
\name{excel_generation}
\alias{excel_generation}
\title{Excel generation}
\usage{
excel_generation(
  gpath = NULL,
  wpath = paste0(tempdir(), "/grnsumm.xlsx"),
  cliquesbool = TRUE,
  ...
)
}
\arguments{
\item{gpath}{path to the graph object ('refinedsumm.rds'). (RDS format)}

\item{wpath}{writing path, where the excel file will be saved. (Default: temp directory)}

\item{cliquesbool}{indicating if cliques method should be added to the summary table. (Default: TRUE)}

\item{...}{every argument you should pass to generatecliques() in case cliquesbool is TRUE.}
}
\value{
Excel containing the mentioned parameters.
}
\description{
From input Gene Regulatory Network, an excel is generated containing a table with
driver to target phenotype dependent relationships. Brief summary containing
drivers normalized xor sum, which is the ratio between drivers-targets connections
present only in one of both phenotypes over all possible connections, and cliques, which are driver genes
that are highly correlated (over a user-decision threshold), and may have been lost during
the fitting process of the LINKER method.
}
\examples{

## For this example, we are going to use a generated 'refinedsumm.rds' from the su2c dataset
## (see vignette of TraRe for more information), which is at the external data folder
## of this package.

gpath <- paste0(system.file('extdata',package='TraRe'),'/refinedsumm.rds')
wpath <- paste0(system.file('extdata',package='TraRe'),'/grnsumm.xlsx')

## We are going to use the drivers dataset from the external data folder as well.
## For more information about generatecliques() please check the help page of it.

dataset<-readRDS(paste0(system.file('extdata',package='TraRe')
                 ,'/tfs_linker_example_eg.rds'))
excel_generation(gpath=gpath,wpath=wpath,dataset=dataset)


}
