% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliques_generatecliques.R
\name{generatecliques}
\alias{generatecliques}
\title{Cliques generation and plot displaying.}
\usage{
generatecliques(
  dataset,
  method,
  correlationth,
  sparsecorrmatrix,
  numcliques,
  mandatorygenes,
  selection = 1
)
}
\arguments{
\item{dataset}{input expression file with genes as rows and samples as columns.}

\item{method}{method to use in the correlation matrix (see stats:cor).}

\item{correlationth}{threshold to consider edge exists.}

\item{sparsecorrmatrix}{boolean variable specifying whether to fix values below threshold to 0 or not.}

\item{numcliques}{number of cliques to be generated.}

\item{mandatorygenes}{array of gene names which is mandatory to include in the returned cliques.}

\item{selection}{integer selecting method.}
}
\value{
List containing the plot generated and a list with all the generated cliques.
}
\description{
Cliques generation and plot displaying of the selected method.
}
\examples{
   \dontrun{
   mat <- matrix(c(0,0,0,0,0,1,0,1,0),3,3)
   g <- graph_from_adjacency_matrix(mat)
   cli <- max_cliques(g)
   tittle <- "Maximizing genes/cliques"
   L <-  350
   numcliques <- 'All'
   mandatorygenes <- c()

   generatecliques(cli,tittle,L,numcliques,mandatorygenes)
   }
}
