% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliques_generategraph.R
\name{generategraph}
\alias{generategraph}
\title{Graph generation from genes correlation adjacency matrix.}
\usage{
generategraph(correlationth, sparsecorrmatrix, pdoutput)
}
\arguments{
\item{correlationth}{threshold to consider edge exists.}

\item{sparsecorrmatrix}{boolean variable specifying whether to fix values below threshold to 0 or not.}

\item{pdoutput}{output from preparedata().}
}
\value{
List containing an igraph object, containing all the genes which have at least one edge as vertex, and
an array of all the genes which do not have correlation values above the threshold.
}
\description{
Graph generation from genes correlation, threshold dependant, adjacency matrix.
}
\examples{
   \dontrun{

   #we load the desired dataset.
   dataset <- readRDS(".../dataset.rds")

   #we select pearson correlation.
   method <- "pearson"

   pdoutput <- preparedata(dataset,method)
   foo <- generategraph(correlationth = 0.6, sparsecorrmatrix = T,pdoutput)
   }
}
