% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_from_graph.R
\name{html_from_graph}
\alias{html_from_graph}
\title{Html generation}
\usage{
html_from_graph(
  gpath = NULL,
  wpath = paste0(tempdir()),
  user_mode = TRUE,
  cliquesbool = TRUE,
  ...
)
}
\arguments{
\item{gpath}{path to the graph object ('refinedsumm.rds'). (RDS format)}

\item{wpath}{writing path, where the html and txts file will be saved. (Default: temp directory)}

\item{user_mode}{boolean indicating if this function is called from user or internaly. (Default: TRUE)}

\item{cliquesbool}{indicating if cliques method should be added to the summary table. (Default: TRUE)}

\item{...}{every argument you should pass to generatecliques() in case cliquesbool is TRUE.}
}
\value{
Html and txts containing the mentioned files.
}
\description{
From input Gene Regulatory Network, an html is generated containing a table with
driver to target phenotype dependent relationships. It is a brief summary containing
drivers normalized xor sum, which is the ratio between drivers-targets connections
present only in one of both phenotypes over all possible connections, and cliques, which are driver genes
that are highly correlated (over a user-decision threshold), and may have been lost during
the fitting process of the LINKER method.
}
\details{
This functions takes place inside the runrewiring function, so it is not recommended to run it on its own.
}
\examples{

## For this example, we are going to use a generated 'refinedsumm.rds' from the su2c dataset
## (see vignette of TraRe for more information), which is at the external data folder
## of this package.

gpath <- paste0(system.file('extdata',package='TraRe'),'/refinedsumm.rds')
wpath <- paste0(system.file('extdata',package='TraRe'))

## We are going to use the drivers dataset from the external data folder as well.
## For more information about generatecliques() please check the help page of it.

dataset<-readRDS(paste0(system.file('extdata',package='TraRe')
                 ,'/tfs_linker_example_eg.rds'))
html_from_graph(gpath=gpath,wpath=wpath,dataset=dataset)


}
