% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linker_runphasetwo.R
\name{LINKER_compute_modules_graph}
\alias{LINKER_compute_modules_graph}
\title{Phase II : bipartitive graphs generation}
\usage{
LINKER_compute_modules_graph(modules, Data, mode = "VBSR", alpha = 1 - 1e-06)
}
\arguments{
\item{modules}{Modules obtained from the phase I linker output.}

\item{Data}{lognorm gene expression matrix.}

\item{mode}{Model to use when stablishing the relation between drivers and targets for the graphs generation. By default, VSBR,}

\item{alpha}{alpha parameter if a LASSO model is chosen.}
}
\value{
igraph object containing the related drivers and targets in the form of a bipartitive graph.
}
\description{
Run second phase of the linker method where a bipartitive graph is generated from the phase I output.
This functions takes place inside the linkerrun function, so it is not recommended to run it on its own.
}
