% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linker_html_functions.R
\name{linker_summarize_rungraphs}
\alias{linker_summarize_rungraphs}
\alias{linker_create_index_page}
\alias{genetable_summary}
\alias{edgeinfo_from_graphs}
\title{Generate summary}
\usage{
linker_summarize_rungraphs(
  rungraphs = NULL,
  iso_table = NULL,
  weighted_chip_evidence = NULL,
  graphstr = "data",
  htmlinfo = list(htmldir = "html/", indexpath = "index.html", txtstr = "txts/"),
  runinfo = list(nregs = "1", ntargets = "5", nboots = "10")
)

linker_create_index_page(
  outdir = "./",
  runtag = "run",
  codedir = "./",
  indexpath = "index.html",
  glossarypath = "glossary.html",
  imgstr = "imgs/",
  txtstr = "txts/"
)

genetable_summary(
  filter,
  filterNeigh,
  tabletype,
  nboots = 10,
  minsupport = 4,
  edgesinfo,
  graphstr,
  htmlinfo
)

edgeinfo_from_graphs(rungraphs, iso_table, weighted_chip_evidence)
}
\arguments{
\item{rungraphs}{example of description}

\item{iso_table}{example of description}

\item{weighted_chip_evidence}{example of description}

\item{graphstr}{example of description}

\item{htmlinfo}{example of description}

\item{runinfo}{example of description}

\item{outdir}{example of description}

\item{runtag}{example of description}

\item{codedir}{example of description}

\item{indexpath}{example of description}

\item{glossarypath}{example of description}

\item{imgstr}{example of description}

\item{txtstr}{example of description}

\item{filter}{example of description}

\item{filterNeigh}{example of description}

\item{tabletype}{example of description}

\item{nboots}{example of description}

\item{minsupport}{example of description}

\item{edgesinfo}{example of description}
}
\value{
summaryfinal, a data frame that contains for each module and graph
the number of graph edges at each possible level of support and the
percentage of cumulative edges with each type of ChIP evidence
}
\description{
Given results of Linker runs and other annotations, builds a summary in order to
afterwards
html website that summarizes graph edges by support and ChIP evidence (if
provided).
}
