% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linker_runlinkernet.R
\name{NET_run}
\alias{NET_run}
\alias{NET_compute_graph_all_LASSO1se}
\alias{NET_compute_graph_all_LASSOmin}
\alias{NET_compute_graph_all_LM}
\alias{NET_compute_graph_all_VBSR}
\title{GRN inference via selected model.}
\usage{
NET_run(
  lognorm_est_counts,
  target_filtered_idx,
  regulator_filtered_idx,
  graph_mode = c("VBSR", "LASSOmin", "LASSO1se", "LM"),
  FDR = 0.05,
  NrCores = 30
)

NET_compute_graph_all_LASSO1se(
  Data,
  regulators_idx,
  target_idx,
  alpha = 1 - 1e-06
)

NET_compute_graph_all_LASSOmin(
  Data,
  regulators_idx,
  target_idx,
  alpha = 1 - 1e-06
)

NET_compute_graph_all_LM(Data, regulators_idx, target_idx)

NET_compute_graph_all_VBSR(Data, regulators_idx, target_idx)
}
\arguments{
\item{lognorm_est_counts}{Matrix of log-normalized estimated counts of the gene expression data (Nr Genes x Nr samples)}

\item{target_filtered_idx}{Index of the target genes on the lognorm_est_counts matrix.}

\item{regulator_filtered_idx}{Index of the regulatory genes on the lognorm_est_counts matrix.}

\item{graph_mode}{Chosen method(s) to generate the edges in the bipartite graph. The available options are "VBSR", "LASSOmin", "LASSO1se" and "LM". By default, all methods are chosen.}

\item{FDR}{The False Discovery Rate correction used for the enrichment analysis. By default, 0.05.}

\item{NrCores}{Nr of computer cores for the parallel parts of the method. Note that the parallelization is NOT initialized in any of the functions. By default, 30.}

\item{Data}{Matrix of log-normalized estimated counts of the gene expression data (Nr Genes x Nr samples).}

\item{regulators_idx}{Index of the regulatory genes on the lognorm_est_counts matrix.}

\item{target_idx}{Index of the target genes on the lognorm_est_counts matrix.}

\item{alpha}{feature selection parameter.}
}
\value{
List containing the GRN graphs and GEA results.
}
\description{
Gene Regulatory Network inference via model selection. \code{NET_compute_graph_all_LASSO1se()} defines
the statistics of drivers and targets to be the Lasso method, choosing 1 standard error from the minimum RSS.
\code{NET_compute_graph_all_LASSOmin()} uses Lasso method, choosing the minimum RSS point. \code{NET_compute_graph_all_LM()}
uses a linear model and \code{NET_compute_graph_all_VBSR()} uses a Variational Bayes Spike Regression.
}
\examples{
   \dontrun{
   #(we generally will load it. We will assume it contains 1000 genes and 43 samples, i.e 1000x43)
   lognorm_est_counts <- readRDS(file="../geneexpressiondata.rds")

   #suppose lognorm_est_counts first 10 genes are drivers
   regulator_filtered_idx <- c(1,2,3,4,5,6,7,8,9,10)

   #and the rest are targets
   target_filtered_idx <- c(11:1000)
   Gene_set_collections <- readRDS(file=".../gsc.rds")

   #The rest have default values.
   foo <- LINKER_run(lognorm_est_counts,regulator_filtered_idx,
                     target_filtered_idx,Gene_set_collections)
   }
}
