% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliques_preparedata.R
\name{preparedata}
\alias{preparedata}
\title{Prepare data from input dataset.}
\usage{
preparedata(dataset, method)
}
\arguments{
\item{dataset}{input expression file with genes as rows and samples as columns.}

\item{method}{method to use in the correlation matrix (see stats:cor).}
}
\value{
List containing hash table with genes variance and correlation matrix in the form of adjacency matrix (diag=0)
}
\description{
Prepare the correlation matrix and the hash table for future uses.
}
\examples{
  \dontrun{
  #this normally is loaded from outside.
  dataset <- readRDS(file="../dataset.rds")

  foo <- preparedate (dataset)
  }
}
