% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewiring_preparerewiring.R
\name{preparerewiring}
\alias{preparerewiring}
\title{Prepare rewiring data for running the method.}
\usage{
preparerewiring(
  name,
  linker_saved_file,
  expr_matrix_file,
  gene_info_file,
  phenotype_file,
  final_signif_thresh,
  regulator_info_col_name = "regulator",
  phenotype_col_name = "Class",
  phenotype_class_vals_string = "NR,R"
)
}
\arguments{
\item{name}{Desired name of the folder which is generated.}

\item{linker_saved_file}{Output file from linker function path.}

\item{expr_matrix_file}{Lognorm counts of the gene expression matrix path.}

\item{gene_info_file}{path of a two-column file containing genes and 'regulator' boolean variable.}

\item{phenotype_file}{path of a two-column file containing used samples and Responder or No Responder 'Class' (NR,R).}

\item{final_signif_thresh}{Significance threshold for the rewiring method. The lower the threshold, the restrictive the method.}

\item{regulator_info_col_name}{Column name of the gene_info_file. By default, 'regulator'.}

\item{phenotype_col_name}{Column name of the phenotype_file. By default, 'Class'.}

\item{phenotype_class_vals_string}{Boolean terms of the phenotype_file at the 'Class' column. By default, (NR,R).}
}
\value{
Return a list containing linker output file path, expression matrix path, geneinfo file path, phenotype file path,
regulator info column name, the phenotype column name, the phenotype class strings (NR,R by default), initial permutations
parameters, the retest threshold, the retest permutations and the significance threshold.
}
\description{
Prepare all the files neccessary for running the gene regulatory network rewiring method.
}
\examples{
\dontrun{
wd <- "mypath"
linker_saved_file<-c(paste(wd,'rds','su2c.rds',sep="/"))
expr_matrix_file<-c(paste(wd,'expression','su2c_expression_ARSI.txt',sep="/"))
gene_info_file<-c(paste(wd,'geneinfo','su2c_geneinfo.txt',sep="/"))
phenotype_file<-c(paste(wd,'clinical','su2c_clinical_ARSI.txt',sep="/"))
final_signif_thresh<-c(0.051)

rewiringinput<-preparerewiring(name="example",linker_rds_v,expr_mat_v,gene_info_v,
                               phenotype_file_v,final_signif_thresh)
}
}
