% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewiring_gene_level.R
\name{rewiring_gene_level}
\alias{rewiring_gene_level}
\title{Perform the rewiring at the gene level}
\usage{
rewiring_gene_level(
  linker_output,
  TraReObj,
  fpath = "",
  final_signif_thresh = 0.05,
  include_cliques = FALSE,
  ImpTH = 0.05,
  cliquesTH = 0.8,
  nrcores = 3,
  outdir = tempdir()
)
}
\arguments{
\item{linker_output}{Output from LINKER_run function (recommended with 50 bootstraps).}

\item{TraReObj}{the TrareObj generated during preprocessing step before GRN inference.}

\item{fpath}{Desired path for the rewiring file to be generated.}

\item{final_signif_thresh}{Significance threshold for the rewiring method. The lower the threshold, the restrictive the method. Default set to 0.05.}

\item{include_cliques}{Boolean specifying to include cliques in the returned score matrix. Default set to FALSE.}

\item{ImpTH}{Threshold for the refinement of the returned list. Default set to 0.05.}

\item{cliquesTH}{Correlation threshold if include_cliques is set to TRUE. Default set to 0.8.}

\item{nrcores}{Number of cores to run the parallelization within the rewiring test (default: 3).}
}
\value{
Return a matrix containing, for each gene (or genes if include_cliques is set to TRUE) the pvalue and odds ratio
from the hypergeometric test within three categories.
}
\description{
Perform the rewiring test to check if regulators are enriched in rewiring modules using a hypergeometric test.
}
