% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewiring_runrewiring.R
\name{runrewiring}
\alias{runrewiring}
\title{GRN Modules Rewiring Method.}
\usage{
runrewiring(
  ObjectList,
  orig_test_perms = 100,
  retest_thresh = 0.08,
  retest_perms = 1000
)
}
\arguments{
\item{ObjectList}{Output from \code{preparerewiring()}containing some required parameters.}

\item{orig_test_perms}{Initial permutations for first test (default: 100) .}

\item{retest_thresh}{Threshold if a second test is performed (default: 1000) .}

\item{retest_perms}{Permutations if a second test is performed (default: 0.08) .}
}
\value{
It creates a folder (in the working directory) containing the files explained above.
}
\description{
Gene Regulatory Network modules Rewiring method. It performs a permutation test,
(what we call rewiring test) and generates an html report containing a correlation matrix
with the higher scores obtained from the rewiring test. This matrix is shown in the way of a
heatmap, and its sorted by a hierarchical clustering for better interpretation.
}
\examples{

## Lets assume that we have already generated the ObjectList, we will load it from
## the folder containing the examples files. After this, it is all straight forward.
## For this example, we will use the default parameters (`orig_test_perms`,`retest_thresh`,
## `retest_perms`)


objectlist <- readRDS(file=paste0(system.file("extdata",package="TraRe"),
                      '/prepared_rewiring_example.rds'))

\dontrun{
 runrewiring(ObjectList = objectlist)
 }



}
