% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliques_selectmethod.R
\name{selectmethod}
\alias{selectmethod}
\title{Method selection for removing duplicities from maximal_cliques method.}
\usage{
selectmethod(selection, ggoutput, pdoutput)
}
\arguments{
\item{selection}{integer selecting method.}

\item{ggoutput}{output from generategraph()}

\item{pdoutput}{output from preparedata().}
}
\value{
List containing the cliques generated, without duplicities, filtered by the chosen method, the selected method tittle and
the index from which the non-correlated genes start (singleton genes).
}
\description{
Method selection in order to remove duplicities generated from the igraph::max_cliques method.
Four possibilities are available.
}
\examples{
   \dontrun{
   chosenmethod <- 1
   g <- igraph::graph object
   singleton <- c("Gene1","Gene2","Gene3")

   foo <- selectmethod(chosenmethod,g,singleton)
   }
}
