% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewiring_html_functions.R
\name{table2html}
\alias{table2html}
\title{Generate html report.}
\usage{
table2html(resultstable)
}
\arguments{
\item{resultstable}{example of param.}

\item{htmlpagefile}{example of param.}

\item{resultsrelpath}{example of param.}

\item{indexpath}{example of param.}

\item{glossarypath}{example of param.}

\item{mytab}{example of param.}

\item{tabletype}{example of param.}

\item{html_idxs}{example of param.}

\item{html_cols}{example of param.}

\item{filestr}{example of param.}

\item{htmlinfo}{example of param.}

\item{outdir}{example of param.}

\item{runtag}{example of param.}

\item{codedir}{example of param.}

\item{imgstr}{example of param.}
}
\description{
Contain functions to generate an html report from the output modules. \code{table2html()}
creates html summary of a set of linker runs on the same data. \code{write_html_table_page()} place all
pngs on same html page, for run directory. \code{write_tables_all()} create html summary of a set of
linker runs on the same data. \code{create_index_page()} generates the rest of pages of the html report.
}
\examples{
\dontrun{
example of example.
}

}
