% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{TraRe}
\alias{TraRe}
\alias{TraRe-package}
\title{Transcriptional Rewiring package for R.}
\description{
What the package does (one paragraph).
}
\section{Package options}{


The following functions are defined, in use order:

\itemize{

\item \code{generatecliques}: Cliques generation and plot displaying of the selected method.

\item \code{LINKER_run}: GRN generation method.

\item \code{NET_run}: Generate networks in order to plot them.

\item \code{runrewiring}: run rewiring method.

}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ubioinformat/TraRePackage/tree/master}
}

}
\author{
\strong{Maintainer}: Jesus De La Fuente Cedeño \email{jdelafuentec@unav.es} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Mikel Hernaez \email{mhernaez@unav.es} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID}) [copyright holder, thesis advisor]
  \item Charles Blatti \email{blatti@illinois.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID}) [copyright holder]
}

}
\keyword{internal}
