% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linker_runlinker.R
\name{LINKER_run}
\alias{LINKER_run}
\title{GRN inference via selected model.}
\usage{
LINKER_run(
  lognorm_est_counts,
  target_filtered_idx,
  regulator_filtered_idx,
  link_mode = c("VBSR", "LASSOmin", "LASSO1se", "LM"),
  graph_mode = c("VBSR", "LASSOmin", "LASSO1se", "LM"),
  module_rep = "MEAN",
  NrModules = 100,
  corrClustNrIter = 100,
  Nr_bootstraps = 10,
  FDR = 0.05,
  NrCores = 30
)
}
\arguments{
\item{lognorm_est_counts}{Matrix of log-normalized estimated counts of the gene expression data (Nr Genes x Nr samples)}

\item{target_filtered_idx}{Index of the target genes on the lognorm_est_counts matrix.}

\item{regulator_filtered_idx}{Index of the regulatory genes on the lognorm_est_counts matrix.}

\item{link_mode}{Chosen method(s) to link module eigengenes to regulators. The available options are "VBSR", "LASSOmin", "LASSO1se" and "LM". By default, all methods are chosen.}

\item{graph_mode}{Chosen method(s) to generate the edges in the bipartite graph. The available options are "VBSR", "LASSOmin", "LASSO1se" and "LM". By default, all methods are chosen.}

\item{module_rep}{Method selected for use. Default set to LINKER.}

\item{NrModules}{Number of modules that are a priori to be found (note that the final number of modules discovered may differ from this value). By default, 100 modules.}

\item{corrClustNrIter}{output from preparedata(). By default, 100.}

\item{Nr_bootstraps}{Number of bootstrap of Phase I. By default, 10.}

\item{FDR}{The False Discovery Rate correction used for the enrichment analysis. By default, 0.05.}

\item{NrCores}{Nr of computer cores for the parallel parts of the method. Note that the parallelization is NOT initialized in any of the functions. By default, 30.}
}
\value{
List containing the GRN raw results, GRN modules, GRN graphs and GEA results.
}
\description{
Gene Regulatory Network inference via model selection.
}
\examples{
   \dontrun{
   #(we generally will load it. We will assume it contains 1000 genes and 43 samples, i.e 1000x43)
   lognorm_est_counts <- readRDS(file="../geneexpressiondata.rds")

   #suppose lognorm_est_counts first 10 genes are drivers
   regulator_filtered_idx <- c(1,2,3,4,5,6,7,8,9,10)

   #and the rest are targets
   target_filtered_idx <- c(11:1000)
   Gene_set_collections <- readRDS(file=".../gsc.rds")

   #The rest have default values.
   foo <- LINKER_run(lognorm_est_counts,regulator_filtered_idx,
                      target_filtered_idx,Gene_set_collections)
   }
}
