% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linker_runphasetwo.R
\name{LINKER_runPhase2}
\alias{LINKER_runPhase2}
\title{Phase II : bipartitive graphs generation}
\usage{
LINKER_runPhase2(modules, Data, mode = "VBSR", alpha = 1 - 1e-06)
}
\arguments{
\item{modules}{Modules obtained from the phase I linker output.}

\item{Data}{Matrix of log-normalized estimated counts of the gene expression
data (Nr Genes x Nr samples)}

\item{mode}{Chosen method(s) to link module eigengenes to regulators. The available options are
"VBSR", "LASSOmin", "LASSO1se" and "LM". By default, all methods are chosen.}

\item{alpha}{alpha parameter if a LASSO model is chosen.}
}
\value{
igraph object containing the related drivers and targets in the form of a bipartitive graph.
}
\description{
Run second phase of the linker method where a bipartitive graph is generated from the phase I output.
This functions takes place inside the linkerrun function, so it is not recommended to run it on its own.
}
