% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_igraph}
\alias{plot_igraph}
\alias{return_layout}
\alias{return_layout_phenotype}
\title{Plotting functions for Gene Regulatory Network.}
\usage{
plot_igraph(mygraph, mytitle, titlecol, mylayout)

return_layout(regs, targets, namehash)

return_layout_phenotype(regs, targets, varfile)
}
\arguments{
\item{mygraph}{igraph::igraph object returned from \code{NET_run()}.}

\item{mytitle}{Desired tittle.}

\item{titlecol}{Color for the tittle.}

\item{mylayout}{desired layout.}

\item{regs}{regulators name list}

\item{targets}{targets name list}

\item{namehash}{dictionary with genes as keys and transcripts as values.
If there is no transcripts, build the dictionary with genes as keys and values. ({"g1":"g1","g2":"g2"})}

\item{varfile}{two column file containing, gene names as rows,
t-statistic from the differential expression analysis of the desired phenotype column and
a boolean variable for regulator (1) - no regulator (0) column.}
}
\value{
None.
}
\description{
Collection of functions for generating graphs layouts to plot GRN obtained from \code{NET_run()} method.
\code{return_layout()} generates a layout from the graph object returned by \code{NET_run()} and \code{return_layout_phenotype()}
plots targets according to the t-statistic from the differential expression analysis of the desired phenotype.
\code{plot_igraph()} takes in the igraph::igraph object and generated layout and generates plot.
}
\examples{
   \dontrun{
   testNet <- NET_run(lognorm_est_counts, target_filtered_idx,
                      regulator_filtered_idx, Gene_set_Collections,
   graph_mode=c("VBSR"),FDR=0.05,NrCores=3)

   mygraph <- testNet$graphs$VBSR
   mylayout <- return_layout(mygraph,c("Reg1","Reg2","Reg3"),c("Target1","Target2","Target3"))

   plot_igraph(mygraph,"MyTittle","black",mylayout)
   }
}
