% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewiring_preparerewiring.R
\name{preparerewiring}
\alias{preparerewiring}
\title{Prepare rewiring data for running the method.}
\usage{
preparerewiring(
  name = "defaultname",
  linker_output = NULL,
  TraReObj = NULL,
  final_signif_thresh = 0.001,
  orig_test_perms = 100,
  retest_thresh = 0.08,
  retest_perms = 1000,
  use_graphs = TRUE,
  outdir = tempdir(),
  nrcores = 3,
  last_cluster = FALSE
)
}
\arguments{
\item{name}{Desired name of the folder which is generated. The chosen
threshold will be \code{paste()} to the folder's name.}

\item{linker_output}{linker output file.}

\item{TraReObj}{TrareObj generated during preprocessing step.}

\item{final_signif_thresh}{Significance threshold for the rewiring method. The lower the threshold, the restrictive the method.}

\item{orig_test_perms}{Initial permutations for first test (default: 100) .}

\item{retest_thresh}{Threshold if a second test is performed (default: 0.08) .}

\item{retest_perms}{Permutations if a second test is performed (default: 1000) .}

\item{use_graphs}{Boolean specifying the use of graphs as linkeroutput modules. (TRUE by default).}

\item{outdir}{Directory for the output folder to be located (default: tempdir())}

\item{nrcores}{Number of cores to run the parallelization within the rewiring test (default: 3).}

\item{last_cluster}{Boolean specifying whether to include the last_cluster in the rewiring or not. (default: FALSE)}
}
\value{
Return a list containing: LINKER's output, expression matrix, boolean array from phenotype file,
array containing number of c(R,NR) samples, significance threshold and output directory.
}
\description{
Prepare neccessary files for running \code{runrewiring()}
}
\examples{

## Load the linker output
linker_output <- readRDS(paste0(system.file('extdata',package='TraRe'),
                                 '/linker_rewiring_example.rds'))

## Load the phenotype file
phenotype_p <- paste0(system.file('extdata',package='TraRe'),
                     '/phenotype_rewiring_example.txt')
phenotype <- read.delim(phenotype_p, row.names = 1)
colnames(phenotype) <- 'phenotype'

## We will be using the same file we generated for the LINKER_run phase.
## TraReObj <- readRDS(paste0(system.file('extdata',package='TraRe'),'/TraReObj.rds'))

## Add the phenotype if it was not in the SummarizedExperiment when we did the preprocessing.
## TraReObj <- rewiring_add_phenotype(TraReObj, phenotype)

## outdir <- system.file('extdata',package='TraRe')
## prepared <- preparerewiring(name='example', linker_output= linker_output,
##                            TraReObj = TraReObj, 
##                            final_signif_thresh=0.05,
##                            nrcores=1,outdir=outdir)

}
