% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewiring_preparerewiring.R
\name{preparerewiring}
\alias{preparerewiring}
\title{Prepare rewiring data for running the method.}
\usage{
preparerewiring(
  name,
  linker_saved_file,
  expr_matrix_file,
  gene_info_file,
  phenotype_file,
  final_signif_thresh,
  regulator_info_col_name = "regulator",
  phenotype_col_name = "Class",
  phenotype_class_vals_string = "NR,R"
)
}
\arguments{
\item{name}{Desired name of the folder which is generated. The chosen
threshold will be \code{paste()} to the name folder.}

\item{linker_saved_file}{Output file from linker function path.}

\item{expr_matrix_file}{Lognorm counts of the gene expression matrix path.}

\item{gene_info_file}{path of a two-column file containing genes and 'regulator' boolean variable.}

\item{phenotype_file}{path of a two-column file containing used samples and Responder or No Responder 'Class' (NR,R).}

\item{final_signif_thresh}{Significance threshold for the rewiring method. The lower the threshold, the restrictive the method.}

\item{regulator_info_col_name}{Column name of the gene_info_file. By default, 'regulator'.}

\item{phenotype_col_name}{Column name of the phenotype_file. By default, 'Class'.}

\item{phenotype_class_vals_string}{Boolean terms of the phenotype_file at the 'Class' column. By default, (NR,R).}
}
\value{
Return a list containing: LINKER's output, expression matrix, boolean array from phenotype file,
array containing number of c(R,NR) samples, significance threshold and output directory.
}
\description{
Prepare neccessary files for running \code{runrewiring()}
}
\examples{

## We are going to prepare 4 files that we have in the example folder: the output from LINKER, the
## gene expression matrix, the phenotype file and the gene info file. Note that the LINKER
## output is generated from the gene expression matrix. Note: if rewiring across more than 1 dataset
## is desired, paths will be given as arrays (i.e. linker_output <- c(path1,path2))


linker_output <- paste0(system.file("extdata",package="TraRe"),'/linker_rewiring_example.rds')

expr_matrix <- paste0(system.file("extdata",package="TraRe"), '/expression_rewiring_example.txt')

gene_info <- paste0(system.file("extdata",package="TraRe"),'/geneinfo_rewiring_example.txt')

phenotype_info <- paste0(system.file("extdata",package="TraRe"),'/phenotype_rewiring_example.txt')


prepared <- preparerewiring(name="example",linker_output,expr_matrix,gene_info,
                            phenotype_info,final_signif_thresh=0.001)


}
