% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewiring_gene_level.R
\name{rewiring_gene_level}
\alias{rewiring_gene_level}
\title{Perform the rewiring at the gene level}
\usage{
rewiring_gene_level(
  linker_output_p,
  lognorm_est_counts_p,
  gene_info_p,
  phenotype_p,
  fpath = "",
  final_sig_th = 0.05,
  include_cliques = FALSE,
  ImpTH = 0.05,
  cliquesTH = 0.8,
  nrcores = 3
)
}
\arguments{
\item{linker_output_p}{Output file from linker function path. RDS format is required.}

\item{lognorm_est_counts_p}{Lognorm counts of the gene expression matrix path.}

\item{gene_info_p}{Path of a two-column file containing genes and 'regulator' boolean variable.}

\item{phenotype_p}{Path of a two-column file containing used samples and Responder or No Responder 'Class' (NR,R).}

\item{fpath}{Desired path for the rewiring file to be generated.}

\item{final_sig_th}{Significance threshold for the rewiring method. The lower the threshold, the restrictive the method. Default set to 0.05.}

\item{include_cliques}{Boolean specifying to include cliques in the returned score matrix. Default set to FALSE.}

\item{ImpTH}{Threshold for the refinement of the returned list. Default set to 0.05.}

\item{cliquesTH}{Correlation threshold if include_cliques is set to TRUE. Default set to 0.8.}

\item{nrcores}{Number of cores to run the parallelization within the rewiring test (default: 3).}
}
\value{
Return a matrix containing, for each gene (or genes if include_cliques is set to TRUE) the pvalue and odds ratio
from the hypergeometric test within three categories.
}
\description{
Perform the rewiring test to check if regulators are enriched in rewiring modules using a hypergeometric test.
}
