% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewiring_runrewiring.R
\name{runrewiring}
\alias{runrewiring}
\title{GRN modules Rewiring method.}
\usage{
runrewiring(
  ObjectList,
  orig_test_perms = 100,
  retest_thresh = 0.08,
  retest_perms = 1000
)
}
\arguments{
\item{ObjectList}{Output from preparerewiring function.}

\item{orig_test_perms}{Initial permutations for first test (default: 100) .}

\item{retest_thresh}{Threshold if a second test is performed (default: 1000) .}

\item{retest_perms}{Permutations if a second test is performed (default: 0.08) .}
}
\value{
None. Generate an html report containing a hierarchical clustering of the rewired modules,
a correlation matrix containing the rewiring scores of the modules as aheatmap and all the rewired
modules drivers and targets statistical information about the rewiring process.
}
\description{
Gene Regulatory Network modules Rewiring method.
}
\examples{
\dontrun{
rewiringinput<-preparerewiring(name="example",linker_rds_v,expr_mat_v,gene_info_v,
                               phenotype_file_v,final_signif_thresh)
runrewiring(rewiringinput)
}

}
