% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewiring_runrewiring.R
\name{runrewiring}
\alias{runrewiring}
\alias{gen_heatmap}
\alias{rawsummary}
\title{GRN Modules Rewiring Method.}
\usage{
runrewiring(ObjectList)

gen_heatmap(
  ObjectList,
  module_membership_list,
  allstats,
  imgdir,
  outdir,
  indexpageinfo,
  cmp = FALSE,
  modmeth = "",
  i = 0
)

rawsummary(
  indexpageinfo,
  rawrunmoddata,
  rawsumm,
  norm_expr_mat_keep,
  outdir,
  foldername_p,
  modmeth,
  cut = FALSE
)
}
\arguments{
\item{ObjectList}{Output from \code{preparerewiring()}containing some required parameters.}

\item{module_membership_list}{hash list containing the pair module-significance
from the rewiring test.}

\item{allstats}{array containing information about module's rewiring.}

\item{imgdir}{path to image directory.}

\item{outdir}{path to the output directory.}

\item{indexpageinfo}{list containing information about index.html page.}

\item{cmp}{boolean indicating the joint heatmap case.}

\item{modmeth}{method of evaluation from LINKER_run.}

\item{i}{integer containing the dupla being evaluated.}

\item{rawrunmoddata}{list containing regulators and target genes from multiplicity table.}

\item{rawsumm}{list containing generated graph related information and cut variable.}

\item{norm_expr_mat_keep}{gene expression matrix from ObjectList.}

\item{foldername_p}{path to the current foldername.}

\item{cut}{boolean to avoid errors from previous generated graphs.}
}
\value{
It creates a folder (in tempdir() by default) containing the files explained above.
}
\description{
Gene Regulatory Network modules Rewiring method. It performs a permutation test,
(what we call rewiring test) and generates an html report containing a correlation matrix
with the higher scores obtained from the rewiring test. This matrix is shown in the way of a
heatmap, and its sorted by a hierarchical clustering for better interpretation.
}
\examples{

## Lets assume that we have already generated the ObjectList, we will load it from
## the folder containing the examples files. After this, it is all straight forward.

objectlist <- readRDS(file=paste0(system.file('extdata',package='TraRe'),
                      '/prepared_rewiring_example.rds'))


## We are going to create the folder containing
## the graphs, reports, etc, and then we are deleting it.
## If you want to keep it, do not run the last line.

## We are modifying output directory for this example.
objectlist$outdir <- paste(getwd(),'examplefolder',sep='/')

runrewiring(ObjectList = objectlist)
unlink(objectlist$outdir,recursive = TRUE)


}
