% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewiring_runrewiring.R
\name{runrewiring}
\alias{runrewiring}
\title{GRN Modules Rewiring Method.}
\usage{
runrewiring(ObjectList)
}
\arguments{
\item{ObjectList}{Output from \code{preparerewiring()}containing some required parameters.}
}
\value{
It creates a folder (in tempdir() by default) containing the files explained above.
}
\description{
Gene Regulatory Network modules Rewiring method. It performs a permutation test,
(what we call rewiring test) and generates an html report containing a correlation matrix
with the higher scores obtained from the rewiring test. This matrix is shown in the way of a
heatmap, and its sorted by a hierarchical clustering for better interpretation.
}
\examples{

## Lets assume that we have already generated the ObjectList, we will load it from
## the folder containing the examples files. After this, it is all straight forward.

objectlist <- readRDS(file=paste0(system.file('extdata',package='TraRe'),
                      '/prepared_rewiring_example.rds'))


## We are going to create the folder containing
## the graphs, reports, etc, and then we are deleting it.
## If you want to keep it, do not run the last line.

## We are modifying output directory for this example.
objectlist$outdir <- paste(getwd(),'examplefolder',sep='/')

runrewiring(ObjectList = objectlist)
unlink(objectlist$outdir,recursive = TRUE)


}
