% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{TraRe}
\alias{TraRe}
\alias{TraRe-package}
\title{Transcriptional Rewiring package for R.}
\description{
What the package does (one paragraph).
}
\section{Package options}{


The following functions are defined, in use order:

\itemize{

\item \code{preparedata}: Prepare the correlation matrix and the hash table for future uses.

\item \code{generategraph}: Cliques generation and plot displaying of the selected method

\item \code{selectmethod}: Method selection in order to remove duplicities generated from
the igraph::max_cliques method. Four possibilities are available.

\item \code{generatecliques}: Cliques generation and plot displaying of the selected method.

}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ubioinformat/TraRePackage/tree/master}
}

}
\author{
\strong{Maintainer}: Jesus De La Fuente Cedeño \email{jdelafuentec@unav.es} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
