% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{trare_preprocessing}
\alias{trare_preprocessing}
\alias{rewiring_add_phenotype}
\title{Preprocess prior to LINKER_run and run_rewiring method}
\usage{
trare_preprocessing(
  data_matrix,
  geneinfo = NULL,
  nassay = 1,
  low_var_genes_th = 0.25,
  low_var_samples_th = 1,
  verbose = TRUE
)

rewiring_add_phenotype(TraReObj, phenotype_f)
}
\arguments{
\item{data_matrix}{Matrix of log-normalized estimated counts of the gene expression
data (Nr Genes x Nr samples) or SummarizedExperiment object.}

\item{geneinfo}{array of the regulator gene names that are in the specified data matrix.}

\item{nassay}{If SummarizedExperiment object is passed as input to lognorm_est_counts, name of the
assay containing the desired matrix. Default: 1 (first item in assay's list).}

\item{low_var_genes_th}{Perform a filtering to drop out low variance (<th) genes across samples. Default: 0.25}

\item{low_var_samples_th}{Perform a filtering to drop out low variance (<th) samples across genes. Default: 1}

\item{verbose}{whether to show or not messages during function run.}

\item{TraReObj}{TraReObj from trare_preprocesing}

\item{phenotype_f}{file containing samples as rownames and a column call 'phenotype' containing the labels}
}
\value{
TraReObj containing preprocessed input matrix
}
\description{
Preprocess data prior to GRN inference process.  Help them for more information.
}
\examples{
   ##

}
